CHANGE: Moved recursive calls to iterative
Since Python doesn't implement TCO, recursive calls can be very bad.
This is also seen when we try to run the algorithm with a lot cuts, we
get a crash due to the recursion limit. We have removed the easily
TCO-optimizable calls with a while loop. A few recursive calls stay
(precisely those at splitting nodes).